Refactor all interaction between Horta viewer and movie making API into new HortaMovieSource class.
Former-commit-id: 957eb7f44aed62567c3e15962ed2ece51dcfda2e

+++ ConsoleApplication/HortaTracer/src/org/janelia/horta/movie/SaveFramesPanel.java
@@ -377,7 +377,7 @@ public class SaveFramesPanel extends JPanel
                         
                         // TODO: check for "Cancel" operation
                         
-                        updateProgressBar((int)Math.round(100 * progressRatio));                
+                        updateProgressBar(Math.round(100 * progressRatio));                
                     }
                     
                     // TODO: post run sanity checks
